feat: add ExerciseListPage component and update routing for exercises
