+++ apps/web/src/feature/exercise/index.ts
@@ -1,0 +1,3 @@
+export { ExerciseListPage } from "./components/ExerciseListPage";
+export { ExerciseForm } from "./components/ExerciseForm";
+export * from "./types";
