+++ apps/web/tsconfig.app.json
@@ -33,7 +33,8 @@
       "@services/*": ["src/services/*"],
       "@assets/*": ["src/assets/*"],
       "@packages/*": ["../../packages/*"],
-      "@store": ["src/store/rootStore.ts"]
+      "@store": ["src/store/rootStore.ts"],
+      "@feature/*": ["src/feature/*"]
     }
   },
   "include": ["src"]
